modularize into calculate single post score
